requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.40
python-dotenv==1.1.0
langchain-google-genai==2.1.2
langchain-community==0.3.20
openevals==0.0.17